+++ packages/browser-wallet/src/popup/pages/AddWeb3IdCredential/i18n/en.ts
@@ -7,7 +7,11 @@ const t = {
             'We are unable to add the web3Id credential to the wallet due to the following issue, please report this to the issuer:',
         metadata: 'We are unable to load the metadata for the credential.',
         schema: 'We are unable to load the schema specification for the credential.',
-        attribute: 'The received credential is missing one or more required attributes ({{ attributeKeys }})',
+        attribute: {
+            required: 'The received credential is missing one or more required attributes ({{ attributeKeys }})',
+            additional:
+                'The attribute with key [{{ credentialAttribute }}] is not available in the list of schema attributes: [{{ schemaAttributes }}]',
+        },
         localization: 'Failed to get localization',
     },
 };