+++ packages/browser-wallet-api/src/wallet-api.ts
@@ -29,7 +29,8 @@ import {
 import EventEmitter from 'events';
 import type { JsonRpcRequest } from '@concordium/common-sdk/lib/providers/provider';
 import { IdProofOutput, IdStatement } from '@concordium/common-sdk/lib/idProofTypes';
-import { CredentialStatements, VerifiablePresentation } from '@concordium/common-sdk/lib/web3ProofTypes';
+import { CredentialStatements } from '@concordium/common-sdk/lib/web3ProofTypes';
+import { VerifiablePresentation } from '@concordium/common-sdk/lib/types/VerifiablePresentation';
 import { ConcordiumGRPCClient } from '@concordium/common-sdk/lib/GRPCClient';
 import JSONBig from 'json-bigint';
 import { stringify } from './util';