+++ packages/browser-wallet/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.1.3
 
 ### Changed
 
@@ -11,6 +11,7 @@
 -   An issue where the import window would fail to open.
 -   Updated the JSON schema for the verifiable credential schema validation, so that invalid schemas are rejected.
 -   An issue where a verifiable with the `NotActivated` status would show as `Pending`.
+-   Enable validation of veriable presentation requests before opening the popup window.
 -   An issue that allowed empty credential statements to be accepted by the wallet-api.
 -   An issue where the wallet allowed for requests adding credentials with more attributes than listed in the schema.